### 2019-03-10 #1312 @whedon
[REVIEW]: Rclean: A Tool for Writing Cleaner, More Transparent Code
**Submitting author:** @MKLau (<a href="http://orcid.org/0000-0003-3758-2406">Matthew Lau</a>)
**Repository:** <a href="https://github.com/ropensci/Rclean" target ="_blank">https://github.com/ropensci/Rclean</a>
**Version:** v1.1.8
