### 2019-03-10 #1312 @labarba
@benmarwick [asked](https://github.com/openjournals/joss-reviews/issues/1242#issuecomment-471356805) in the Pre-review issue whether submitting to Ropensci had been considered. Comments, @MKLau?
