### 2019-03-15 #1312 @MKLau
@labarba @benmarwick
No we hadn't thought to submit to Ropensci but I am familiar with the work there and would be interested in submitting there.

Per "Has there been any discussion about whether this pkg might be suitable for @ropensci onboarding? Perhaps @noamross or @maelle could advise if this pkg is suitable? It it passes that, then it's auto-submitted here, if I understand correctly."

The on-boarding process sounds very useful. Also, I am familiar with Noam Ross's work and would think that he would be suitable to review.

I'm not quite sure that I fully understand the suggestion. Should a submission to ROpenSci be done in addition to (e.g. for on-boarding) or in-lieu of submitting to JOSS?
