### 2019-03-15 #1312 @labarba
We have an agreement with ROpenSci where if your package goes through their review, it gets fast-tracked to a publication in JOSS with minor editorial checks.
