### 2019-03-16 #1312 @labarba
@cboettig, @benmarwick — Thank you for agreeing to review this JOSS submission. The review is now _paused_, while the author investigates going the ROpenSci route. Stay tuned!
