### 2019-05-13 #1312 @MKLau
@labarba Yes, I'm almost done with a couple of pre-submission edits based on the ROpenSci guidelines. Shooting to submit before the end of this month.
