### 2019-05-13 #1312 @maelle
@MKLau before submission to rOpenSci you can open a [pre-submission inquiry](https://github.com/ropensci/software-review/issues/new/choose) so that the editors might assess whether your package is in scope. Thank you! :smile_cat:
