### 2019-05-14 #1312 @MKLau
Submitted a presub inquiry see #[300](https://github.com/ropensci/software-review/issues/300).
