### 2019-06-28 #1312 @MKLau
Hi Daniel, sorry for the slow reply I've been traveling and just saw this post. Not much happened last month, but shouldn't be more than a week to get things finished for submission. There are a few more functions that need some tests and the vignette needs to be added:  https://github.com/ProvTools/Rclean/projects/4.
