### 2019-07-15 #1312 @MKLau
Hi @danielskatz @labarba @benmarwick, the package review should be good to start. I made an inquiry over in pre-submission thread https://github.com/ropensci/software-review/issues/300 to see what the best way would be to proceed. One of you might have a thought as well. Should I just re-open the closed pre-submission thread and re-label it? Or, should I start a new issue?

Thanks!

@MKLau
