### 2019-07-15 #1312 @labarba
I don't know how you submit to rOpenSci, but that's what you need to do next. Once your software is reviewed there, it will be fast-tracked in JOSS.
