### 2019-07-15 #1312 @cboettig
@labarba @MKLau To submit to rOpenSci, simply open an issue in https://github.com/ropensci/software-review/ as described in the README there.
