### 2019-10-29 #1312 @maelle
The package is under review at rOpenSci https://github.com/ropensci/software-review/issues/327 cc handling editor @annakrystalli
