### 2020-02-10 #1312 @danielskatz
@maelle & @MKLau - Has the rOpenSci package now been accepted? I see that the status is 6/approved, but am not sure if there's another step after that.
