### 2020-02-10 #1312 @MKLau
Hi @danielskatz ,

It has been accepted. I’m currently finishing the transfer to ROpenSci org and revisions of the manuscript.
