### 2020-02-11 #1312 @danielskatz
@MKLau - please let us know when the paper is complete, so we can proceed to accept this in JOSS.  We don't need to wait for the transfer to rOpenSci, but we do need the final paper.
