### 2020-02-12 #1312 @MKLau
Hi @danielskatz , just finished sorting out a couple of issues with Travis and Zenodo tracking.

The manuscript is now good to go for review, you can find it [here](https://github.com/ropensci/Rclean/joss/paper.Rmd).
