### 2020-02-13 #1312 @danielskatz
@MKLau - we need the paper to be a .md file here, with a .bib file, as shown in https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography  We will then build it with ```@whedon generate pdf``` (you can do this too to check)
