### 2020-02-13 #1312 @danielskatz
@MKLau - please fix the problems in the .md and/or .bib files, perhaps updating them based on recent changes made for the rOpenSci process
