### 2020-02-14 #1312 @danielskatz
👋 @cboettig, @benmarwick - as this is an rOpenSci paper, we don't need an independent JOSS review of the software, so your help will not be called on - also, sorry for what might have been too many notifications...
