### 2020-02-14 #1312 @danielskatz
Also, please merge the changes in https://github.com/ropensci/Rclean/pull/202
