### 2020-02-14 #1312 @MKLau
Thanks @danielskatz for the edits.

I've reviewed and accepted your pull request and tagged as v1.1.8.

Here's the DOI and link to Zenodo.

10.5281/zenodo.3665732

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3665732.svg)](https://doi.org/10.5281/zenodo.3665732)
