### 2020-02-15 #1312 @danielskatz
There are also some changes needed in the bib file - please merge https://github.com/ropensci/Rclean/pull/203
