### 2020-02-15 #1312 @danielskatz
👋 @openjournals/dev - note the second author's name has gotten parsed incorrectly in the XML - should I accept this and then ask you to fix this manually?  Or can we do something before that?
