### 2020-02-15 #1312 @MKLau
@danielskatz @arfon

If the multiple middle initials is a problem, he can be cited as Thomas Pasquier, as in this [paper](https://tfjmp.org/files/publications/scidata-2017.pdf).
