### 2020-02-15 #1312 @danielskatz
It's not a problem, we can handle it.  However, we are waiting on
> There are also some changes needed in the bib file - please merge ropensci/Rclean#203
