### 2020-02-15 #1312 @MKLau
@danielskatz Great, pull request accepted. Thanks for those fixes.
