### 2020-02-17 #1312 @arfon
> 👋 @arfon - over to you to fix the 2nd author in the metadata (xml file), then to close this

Done.
